chore: integrate OpenRouter AI SDK and update AI configuration management
